+++ core/routes/collection.ts
@@ -1,5 +1,5 @@
 import { changesToData, getViews, prepareField } from "../utils/data.ts";
-import { getPath, normalizeName } from "../utils/path.ts";
+import { getLanguageCode, getPath, normalizeName } from "../utils/path.ts";
 import { posix } from "../../deps/std.ts";
 import { render } from "../../deps/vento.ts";
 
@@ -342,24 +342,4 @@ function get(c: Context) {
     options,
     versioning,
   };
-}
-
-function getLanguageCode(path: string): string {
-  const ext = posix.extname(path).toLowerCase();
-  switch (ext) {
-    case ".md":
-      return "Markdown";
-    case ".yml":
-      return "YAML";
-    case ".json":
-      return "JSON";
-    case ".css":
-      return "CSS";
-    case ".js":
-      return "JavaScript";
-    case ".ts":
-      return "TypeScript";
-    default:
-      return "HTML";
-  }
 }