+++ core/routes/document.ts
@@ -1,4 +1,4 @@
-import { getPath } from "../utils/path.ts";
+import { getLanguageCode, getPath } from "../utils/path.ts";
 import { changesToData, getViews, prepareField } from "../utils/data.ts";
 import { render } from "../../deps/vento.ts";
 
@@ -44,6 +44,54 @@ export default function (app: Hono) {
 
       await document.write(changesToData(body), options, true);
       return c.redirect(getPath(options.basePath, "document", document.name));
+    });
+
+  app
+    .get("/document/code/:document", async (c: Context) => {
+      const { document, versioning } = get(c);
+
+      if (!document) {
+        return c.notFound();
+      }
+
+      const code = await document.readText();
+      const fields = [{
+        tag: "f-code",
+        name: "code",
+        label: "Code",
+        type: "code",
+        attributes: {
+          data: {
+            language: getLanguageCode(document.name),
+          },
+        },
+      }];
+      const data = { code };
+
+      try {
+        return c.render(
+          await render("document/code.vto", {
+            fields,
+            data,
+            document,
+            version: versioning?.current(),
+          }),
+        );
+      } catch (e) {
+        console.error(e);
+        return c.notFound();
+      }
+    })
+    .post(async (c: Context) => {
+      const { options, document } = get(c);
+
+      const body = await c.req.parseBody();
+      const code = body["changes.code"] as string | undefined;
+      document.writeText(code ?? "");
+
+      return c.redirect(
+        getPath(options.basePath, "document", "code", document.name),
+      );
     });
 }