+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.12.3] - Unreleased
 ### Added
 - Support to configure the path in `GitHub::create()`.
+- Allow to edit documents in code mode.
 - Allow to customize the language of `code` field with `attributes.data.language`.
 - Support for front matter in the `code` field.
 - When the parsing of a collection's document fails, show the error and allow to edit the raw code.